fixed up judge view page /judge
