+++ components/other/LineGraph.tsx
@@ -1,0 +1,28 @@
+import { Line } from 'react-chartjs-2';
+import {
+  Chart as ChartJS,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend,
+} from 'chart.js';
+import { FunctionComponent } from 'react';
+
+ChartJS.register(
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend
+);
+
+export const LineGraph: FunctionComponent<{data: any}> = ({data}) => {
+    return (
+        <Line data={data} />
+    )
+}
