+++ types/Evaluation.ts
@@ -1,5 +1,6 @@
 export type Evaluation = {
     tournamentName: string,
+    date: Date | string,
     roundName: string, // e.g., Round 1 Flight A etc.
     isPrelim: boolean,
     isImprovement: boolean,
